% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get data by making GET request to obs db API}
\usage{
getData(
  base_url,
  auth_key,
  dataset_id,
  stationId,
  phenomenonId,
  offeringId,
  unitId,
  ...
)
}
\description{
Get data by making GET request to obs db API
}
\keyword{internal}
