% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDatasets.R
\name{getDatasets}
\alias{getDatasets}
\title{Download datasets}
\usage{
getDatasets(df, startDate = NULL, endDate = NULL)
}
\arguments{
\item{df}{A data.frame with rows corresponding to datasets to download. Should be generated from \code{\link{viewDatasets}}. Must contain columns `stationID`, `phenomenomId`, `InstrumentId` and `UnitId`}

\item{startDate}{Start date for observations. Either a date or string with format 'YYYY-MM-DD' or 'YYYY-MM-DD HH-MM-SS'. Defaults to NULL.}

\item{endDate}{End date for observations. Either a date or string with format 'YYYY-MM-DD' or 'YYYY-MM-DD HH-MM-SS'. Defaults to NULL.}
}
\value{
A data.frame with observation data and associated details
}
\description{
Requests and downloads datasets from the SAEON observations database
}
\examples{
\dontrun{
#get dataframe with available dataset and filter based on search criteria

df <- viewDatasets() \%>\%
      filter(siteName == 'Constantiaberg') \%>\%
      filter(description == 'Air Temperature - Daily Minimum - Degrees Celsius')

#fetch those datasets for 2020

obs <- getDatasets(df,startDate='2020-01-01',endDate='2020-12-31')
}

}
