% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postData.R
\name{postData}
\alias{postData}
\title{Get data by making POST request to obs db API}
\usage{
postData(
  base_url,
  auth_key,
  dataset_id,
  stationId,
  phenomenonId,
  offeringId,
  unitId,
  startDate,
  endDate,
  ...
)
}
\description{
Get data by making POST request to obs db API
}
\keyword{internal}
