% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewDatasets.R
\name{viewDatasets}
\alias{viewDatasets}
\title{View datasets}
\usage{
viewDatasets(extent = NULL, spatial = FALSE)
}
\arguments{
\item{extent}{A spatial object of class `sf` object giving the extent over to search for datasets. Defaults to NULL}

\item{spatial}{Should the result be returned as a sf object? Defaults to FALSE}
}
\value{
A data.frame with available datasets and relevant details
}
\description{
Lists datasets available in the SAEON observations database
}
\examples{
\dontrun{
viewDatasets()
}

}
